refactor about solved challenge, rebuild filter method

+++ src/main/java/com/example/codelabsvc/entity/TestCase.java
@@ -36,7 +36,7 @@ public class TestCase {
 
     private Integer compilationDuration;
 
-    private Integer averageExecutionDuration;
+    private Double averageExecutionDuration;
 
     private Integer timeLimit;
